+++ Character/Class/Priest/Priest.cpp
@@ -70,7 +70,7 @@ unsigned Priest::get_spirit_modifier() const {
 }
 
 unsigned Priest::get_mp5_from_spirit() const {
-    return 13 + static_cast<unsigned>(round(cstats->get_spirit() / 4));
+    return 13 + static_cast<unsigned>(round(static_cast<double>(cstats->get_spirit()) / 4));
 }
 
 double Priest::get_agi_needed_for_one_percent_phys_crit() const {
